refactor: Implement socket controller finishExam
- move CRUD logic related to exam deletion and data validation to new controller factory function `deleteExam` from `handleExams.js` to `examControllers.js`.
- add additional validation and support for factory models and error handling in `deleteExam` in `examControllers.js`.
- change handler function from `deleteExam` to `handleDeleteExam` in `handleExams.js` and update the instances of handler name reference in  `manageExams.js`.
