+++ backend/controller/socket/handlers/handleExams.js
@@ -4,7 +4,11 @@ import Quiz from "../../../models/quizzes/quizModel.js";
 import Classroom from "../../../models/classrooms/classroomModel.js";
 import Score from "../../../models/quizzes/scoreModel.js";
 import { handleNewNotification } from "./handleNotifications.js";
-import { createExam, updateExam } from "../controllers/examControllers.js";
+import {
+  createExam,
+  finishExam,
+  updateExam,
+} from "../controllers/examControllers.js";
 
 export const handleCreateExam = async (context, data) => {
   try {
@@ -49,12 +53,14 @@ export const handleUpdateExam = async (context, data) => {
   }
 };
 
-export const finishExam = async (context, data) => {
+export const handleFinishExam = async (context, data) => {
   try {
-    const examEndPayload = await createExamPayload(data);
+    const models = { Exam, Quiz, Score };
+
+    const examEndPayload = await finishExam(models, data);
 
     if (!examEndPayload.success) {
-      throw new Error(examEndPayload.message);
+      throw new Error(examEndPayload?.message || "Unable to finish exam");
     }
 
     const ongoingJob = schedule.scheduledJobs[data?.senderId];
@@ -63,7 +69,7 @@ export const finishExam = async (context, data) => {
       ongoingJob.cancel();
     }
 
-    context.emitEvent("sender", "exam finished", examEndPayload.examPayload);
+    context.emitEvent("sender", "exam finished", examEndPayload?.examPayload);
 
     const notificationData = {
       senderId: data?.senderId,
@@ -75,99 +81,5 @@ export const finishExam = async (context, data) => {
     await handleNewNotification(context, notificationData);
   } catch (error) {
     console.error(error.message);
-  }
-};
-
-export const createExamPayload = async (data) => {
-  try {
-    const { senderId, quizId } = data;
-
-    const examFound = await Exam.findOne({ studentId: senderId });
-
-    if (!examFound) {
-      throw new Error("Exam not found");
-    }
-
-    const examQuestions = examFound?.shuffledQuestions;
-
-    const examIsValid = examFound?.examFinish?.getTime() - Date.now() > 0;
-
-    if (!examIsValid) {
-      throw new Error("Exam not found");
-    }
-
-    await Exam.findByIdAndUpdate(
-      examFound?._id,
-      { isInProgress: false },
-      { new: true }
-    );
-
-    const quizFound = await Quiz.findOne({ _id: quizId });
-
-    if (!quizFound) {
-      throw new Error("Cannot evaluate exam, no matching quiz found");
-    }
-
-    let currentScore = 0;
-
-    let userChoices = [];
-
-    for (let i = 0; i < quizFound.questions.length; i++) {
-      const q = quizFound.questions[i];
-
-      if (q.answer === examFound.answers[i]) {
-        currentScore += 1;
-      }
-
-      userChoices[i] = {
-        userAnswer: examFound?.answers?.[i],
-        correctAnswer: q.answer,
-      };
-    }
-
-    const scoreFound = await Score.findOne({
-      user: senderId,
-      "quiz.quizId": quizFound?._id,
-    });
-
-    let scorePayload;
-
-    if (scoreFound) {
-      scorePayload = await Score.findByIdAndUpdate(
-        scoreFound?._id,
-        {
-          $set: {
-            "examFeedback.userChoices": userChoices,
-            "examFeedback.randomizedQuestions": examQuestions,
-            highScore: Math.max(scoreFound.highScore, currentScore),
-            latestScore: currentScore,
-          },
-        },
-        { new: true }
-      );
-    } else {
-      scorePayload = await new Score({
-        user: senderId,
-        quiz: {
-          quizId: quizFound._id,
-          quizTitle: quizFound.title,
-        },
-        examFeedback: {
-          userChoices,
-          randomizedQuestions: examQuestions,
-        },
-        highScore: currentScore,
-        latestScore: currentScore,
-      }).save();
-    }
-
-    const examPayload = { examId: examFound?._id, scorePayload };
-
-    await Exam.findByIdAndDelete({ _id: examFound?._id });
-
-    return { success: true, examPayload };
-  } catch (error) {
-    console.error(`Error creating exam payload: ${error.message}`);
-    return { success: false, message: `Error finishing exam ${error.message}` };
   }
 };
